added AppliedFiltersSection
