+++ projects/ui/src/Components/Apis/ApisTab/ApisFilter.tsx
@@ -3,17 +3,14 @@ import { useContext } from "react";
 import { Icon } from "../../../Assets/Icons";
 import { AppContext } from "../../../Context/AppContext";
 import { FilterStyles as Styles } from "../../../Styles/shared/Filters.style";
-import { FilterPair, FilterType } from "../../../Utility/filter-utility";
+import { FilterType } from "../../../Utility/filter-utility";
+import {
+  AppliedFiltersSection,
+  FiltrationProp,
+} from "../../Common/Filters/AppliedFiltersSection";
 import GridListToggle from "../../Common/GridListToggle";
 
-type ApisFiltrationProp = {
-  allFilters: FilterPair[];
-  setAllFilters: (newFiltersList: FilterPair[]) => void;
-  nameFilter: string;
-  setNameFilter: (newNamesList: string) => void;
-};
-
-export function ApisFilter({ filters }: { filters: ApisFiltrationProp }) {
+export function ApisFilter({ filters }: { filters: FiltrationProp }) {
   const { preferGridView, setPreferGridView } = useContext(AppContext);
 
   // const [pairFilter, setPairFilter] = useState<KeyValuePair>({
@@ -77,20 +74,6 @@ export function ApisFilter({ filters }: { filters: ApisFiltrationProp }) {
       ...filters.allFilters,
       { displayName: addedType, type: FilterType.apiType },
     ]);
-  };
-
-  const removeFilter = (filterPair: FilterPair) => {
-    filters.setAllFilters(
-      filters.allFilters.filter(
-        (filter) =>
-          filter.type !== filterPair.type ||
-          filter.displayName !== filterPair.displayName
-      )
-    );
-  };
-
-  const clearAll = () => {
-    filters.setAllFilters([]);
   };
 
   const selectableTypes = [
@@ -175,34 +158,7 @@ export function ApisFilter({ filters }: { filters: ApisFiltrationProp }) {
           isList={!preferGridView}
         />
       </div>
-
-      {filters.allFilters.length > 0 && (
-        <div className="currentFiltersArea">
-          <Styles.ActiveFiltersGrid>
-            {filters.allFilters.map((activeFilter, idx) => (
-              <Styles.ActiveFilter key={idx}>
-                {activeFilter.displayName}
-                <button
-                  className="closingX"
-                  aria-label={`Remove ${activeFilter.displayName} filter`}
-                  onClick={() => removeFilter(activeFilter)}
-                >
-                  <Icon.SmallX />
-                </button>
-              </Styles.ActiveFilter>
-            ))}
-          </Styles.ActiveFiltersGrid>
-          <Styles.ClearAllButton
-            aria-label={`Remove all filters`}
-            onClick={clearAll}
-          >
-            Clear All
-            <span className="closingX">
-              <Icon.SmallX />
-            </span>
-          </Styles.ClearAllButton>
-        </div>
-      )}
+      <AppliedFiltersSection filters={filters} />
     </Styles.FilterArea>
   );
 }
