+++ projects/ui/src/Components/Apps/PageContent/AppsFilter.tsx
@@ -5,20 +5,17 @@ import { Icon } from "../../../Assets/Icons";
 import { AppContext } from "../../../Context/AppContext";
 import { FilterStyles as Styles } from "../../../Styles/shared/Filters.style";
 import { FilterPair, FilterType } from "../../../Utility/filter-utility";
+import {
+  AppliedFiltersSection,
+  FiltrationProp,
+} from "../../Common/Filters/AppliedFiltersSection";
 import GridListToggle from "../../Common/GridListToggle";
-
-type AppsFiltrationProp = {
-  allFilters: FilterPair[];
-  setAllFilters: (newFiltersList: FilterPair[]) => void;
-  nameFilter: string;
-  setNameFilter: (newNamesList: string) => void;
-};
 
 export function AppsFilter({
   filters,
   teams,
 }: {
-  filters: AppsFiltrationProp;
+  filters: FiltrationProp;
   teams: Team[];
 }) {
   const { preferGridView, setPreferGridView } = useContext(AppContext);
@@ -121,33 +118,7 @@ export function AppsFilter({
         />
       </div>
 
-      {filters.allFilters.length > 0 && (
-        <div className="currentFiltersArea">
-          <Styles.ActiveFiltersGrid>
-            {filters.allFilters.map((activeFilter, idx) => (
-              <Styles.ActiveFilter key={idx}>
-                {activeFilter.displayName}
-                <button
-                  className="closingX"
-                  aria-label={`Remove ${activeFilter.displayName} filter`}
-                  onClick={() => removeFilter(activeFilter)}
-                >
-                  <Icon.SmallX />
-                </button>
-              </Styles.ActiveFilter>
-            ))}
-          </Styles.ActiveFiltersGrid>
-          <Styles.ClearAllButton
-            aria-label={`Remove all filters`}
-            onClick={clearAll}
-          >
-            Clear All
-            <span className="closingX">
-              <Icon.SmallX />
-            </span>
-          </Styles.ClearAllButton>
-        </div>
-      )}
+      <AppliedFiltersSection filters={filters} />
     </Styles.FilterArea>
   );
 }
